Add the posix thread to achieve parallelism
Due to each row of pixel is independent calculated,
it is inefficiently use just one thread to calculate the all pixels.

The arugment of thread function only accept the type of  the void* object,
so we wrapped the information needed into a struct,which we can pass into the thread funcion.

Due to race condition, use mutex to protect the index that each thread get and set.

-use multi thread to achieve parallelism.
-wrap the information needed into a structure.
-use mutex